scripts/requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
